Move types to helper redis types file
